Revert "Merge branch 'enab-dev' into backend-dev"
This reverts commit e6a2520f30015c41c94c188cd352922e0da71e0f, reversing
changes made to 614d6a197f119440034a7c34f9e48e2bb5f381d2.
