+++ backend/src/api/controller/article.controller.ts
@@ -43,13 +43,12 @@ export class ArticleController {
 
   // submits articles and sends it to the database
   @Post()
-
   async create(
     @Body() submitArticleDTO: SubmitArticleDTO,
     @Body('email') email: string,
   ) {
     const userEmail = email;
-    return this.ArticleService.create(submitArticleDTO, email);
+    return this.ArticleService.create(submitArticleDTO, userEmail);
   }
 
   // Upload a bibtex file for article submission
