+++ backend/src/api/service/article.service.ts
@@ -287,4 +287,3 @@ export class ArticleService {
     return sort;
   }
 }
-
