+++ frontend/src/components/PopulatedNavBar.tsx
@@ -25,7 +25,9 @@ const PopulatedNavBar = () => {
                 <NavItem dropdown>
                     User Management <IoMdArrowDropdown />
                     <NavDropdown>
-                        <NavItem route="/users">View Users</NavItem>
+                        <NavItem route="/admin/users">View Users</NavItem>
+                        <NavItem route="/admin/permissions">Manage Permissions</NavItem>
+                        <NavItem route="/admin/settings">Settings</NavItem>
                     </NavDropdown>
                 </NavItem>
             )}
