+++ frontend/src/styles/AnalystClaim.module.css
@@ -1,0 +1,155 @@
+/* Container Styles */
+.container {
+  max-width: 1200px;
+  margin: 40px auto;
+  padding: 20px;
+  background-color: #e6f7ff; /* Light bluish background */
+  border-radius: 10px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  display: flex;
+  flex-direction: row;
+  gap: 40px;
+}
+
+/* Headings */
+.heading {
+  text-align: center;
+  color: #34495e;
+  font-size: 32px;
+  font-weight: bold;
+  margin-bottom: 30px;
+}
+
+.subHeading {
+  font-size: 24px;
+  color: #2980b9;
+  margin-bottom: 20px;
+}
+
+/* Error Message */
+.error {
+  color: #e74c3c; /* Red color for errors */
+  text-align: center;
+  margin-bottom: 20px;
+}
+
+/* No Articles Message */
+.noArticles {
+  text-align: center;
+  font-size: 18px;
+  color: #7f8c8d; /* Gray color */
+}
+
+/* Article List Styles */
+.articleListContainer {
+  flex: 1;
+}
+
+.articleList {
+  list-style-type: none;
+  padding: 0;
+}
+
+.articleItem {
+  background-color: #f1f1f1;
+  margin: 10px 0;
+  padding: 15px;
+  border-radius: 5px;
+  cursor: pointer;
+  border: 2px solid transparent;
+  transition: border-color 0.3s ease-in-out;
+}
+
+.articleItem:hover {
+  border-color: #2980b9;
+}
+
+.selected {
+  border-color: #3498db; /* Darker blue when selected */
+}
+
+.articleTitle {
+  font-size: 18px;
+  font-weight: bold;
+}
+
+/* Form Container */
+.formContainer {
+  flex: 1;
+}
+
+/* Claim Info */
+.claimInfo {
+  background-color: #ecf0f1;
+  padding: 15px;
+  border-radius: 5px;
+  margin-bottom: 20px;
+}
+
+.claimInfo p {
+  margin: 5px 0;
+}
+
+/* Form Styles */
+.form {
+  display: flex;
+  flex-direction: column;
+}
+
+.formGroup {
+  margin-bottom: 15px;
+}
+
+.formGroup label {
+  display: block;
+  margin-bottom: 5px;
+  font-weight: bold;
+  color: #2c3e50;
+}
+
+/* Scoped textarea class */
+.textAreaInput {
+  padding: 10px;
+  border: 1px solid #bdc3c7;
+  border-radius: 5px;
+  width: 100%;
+  font-size: 14px;
+  resize: vertical;
+  min-height: 100px;
+}
+
+.textAreaInput:focus {
+  border-color: #3498db; /* Focus color */
+  outline: none;
+}
+
+/* Scoped select field class */
+.selectInput {
+  padding: 10px;
+  border: 1px solid #bdc3c7;
+  border-radius: 5px;
+  width: 100%;
+  font-size: 14px;
+}
+
+.selectInput:focus {
+  border-color: #3498db; /* Focus color */
+  outline: none;
+}
+
+/* Submit Button */
+.submitButton {
+  padding: 12px 20px;
+  font-size: 16px;
+  background-color: #2980b9; /* Blue */
+  color: white;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.3s ease-in-out;
+  align-self: flex-start;
+}
+
+.submitButton:hover {
+  background-color: #21618c; /* Darker blue */
+}
