--- frontend/src/styles/ArticleSearch.module.css
@@ -1,77 +1,0 @@
-/* ArticleResearch.module.css */
-
-.container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 20px;
-  }
-  
-  .heading {
-    font-size: 2em;
-    margin-bottom: 20px;
-  }
-  
-  .filterSection {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    width: 80%;
-    margin-bottom: 20px;
-  }
-  
-  .filterGroup {
-    margin-bottom: 15px;
-  }
-  
-  .filterGroup label {
-    font-weight: bold;
-    margin-right: 10px;
-  }
-  
-  .filterInput {
-    padding: 8px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-  }
-  
-  .searchBox {
-    padding: 8px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    width: 300px;
-  }
-  
-  .error {
-    color: red;
-    margin-bottom: 20px;
-  }
-  
-  .noArticles {
-    font-size: 1.2em;
-    color: #555;
-  }
-  
-  .table {
-    width: 90%;
-    border-collapse: collapse;
-    margin-top: 20px;
-    border: 1px solid #ddd;
-  }
-  
-  .table th,
-  .table td {
-    border: 1px solid #ddd;
-    padding: 12px 15px;
-    text-align: left;
-  }
-  
-  .table th {
-    background-color: #f2f2f2;
-    font-weight: bold;
-  }
-  
-  .table tr:nth-child(even) {
-    background-color: #f9f9f9;
-  }
-
