+++ frontend/src/styles/Moderator.module.css
@@ -1,35 +1,49 @@
-/* Moderator.module.css */
-
+/* Existing styles for container and table */
 .container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
+  max-width: 1200px;
+  margin: 40px auto;
   padding: 20px;
+  background-color: #ffffff; /* White background */
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
 }
 
 .heading {
-  font-size: 2em;
+  text-align: center;
+  margin-bottom: 30px;
+  color: #2c3e50; /* Dark blue-gray */
+  font-size: 32px;
+  font-weight: bold;
+}
+
+.error {
+  color: #e74c3c; /* Red color for errors */
+  text-align: center;
   margin-bottom: 20px;
 }
 
+.noArticles {
+  text-align: center;
+  font-size: 18px;
+  color: #7f8c8d; /* Gray color */
+}
+
 .table {
-  width: 80%;
+  width: 100%;
   border-collapse: collapse;
   margin-top: 20px;
-  border: 1px solid #ddd;
 }
 
 .table th,
 .table td {
-  border: 1px solid #ddd;
   padding: 12px 15px;
+  border-bottom: 1px solid #ddd;
   text-align: left;
 }
 
 .table th {
   background-color: #f2f2f2;
-  font-weight: bold;
+  color: #2c3e50;
 }
 
 .table tr:nth-child(even) {
@@ -41,50 +55,70 @@
   gap: 10px;
 }
 
-.acceptButton {
-  background-color: #4CAF50;
-  color: white;
+/* Button Styles */
+.acceptButton,
+.rejectButton {
+  padding: 8px 12px;
+  margin-right: 5px;
+  font-size: 14px;
   border: none;
-  padding: 10px 20px;
+  border-radius: 5px;
   cursor: pointer;
+}
+
+.acceptButton {
+  background-color: #2ecc71; /* Green */
+  color: #fff;
 }
 
 .rejectButton {
-  background-color: #f44336;
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  cursor: pointer;
+  background-color: #e74c3c; /* Red */
+  color: #fff;
 }
 
-.acceptButton:hover,
+.acceptButton:hover {
+  background-color: #27ae60;
+}
+
 .rejectButton:hover {
-  opacity: 0.8;
+  background-color: #c0392b;
 }
 
+/* Feedback Modal Styles */
 .feedbackModal {
   position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: white;
-  padding: 20px;
-  border: 1px solid #ddd;
-  z-index: 1000;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5); /* Dark overlay */
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 
 .feedbackContent {
-  display: flex;
-  flex-direction: column;
+  background-color: #fff;
+  padding: 20px;
+  width: 500px;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
 }
 
 .feedbackInput {
   width: 100%;
   height: 100px;
-  margin-bottom: 20px;
   padding: 10px;
-  border: 1px solid #ddd;
+  font-size: 14px;
+  border: 1px solid #bdc3c7;
+  border-radius: 5px;
+  margin-bottom: 20px;
+  resize: vertical;
+}
+
+.feedbackInput:focus {
+  border-color: #3498db; /* Focus color */
+  outline: none;
 }
 
 .feedbackActions {
@@ -93,29 +127,27 @@
 }
 
 .submitButton {
-  background-color: #4CAF50;
+  padding: 10px 20px;
+  background-color: #e74c3c; /* Red */
   color: white;
   border: none;
-  padding: 10px 20px;
+  border-radius: 5px;
   cursor: pointer;
+}
+
+.submitButton:hover {
+  background-color: #c0392b; /* Darker red */
 }
 
 .cancelButton {
-  background-color: #f44336;
+  padding: 10px 20px;
+  background-color: #7f8c8d; /* Gray */
   color: white;
   border: none;
-  padding: 10px 20px;
+  border-radius: 5px;
   cursor: pointer;
 }
 
-.noArticles {
-  text-align: center;
-  margin-top: 20px;
-}
-
-.footer {
-  margin-top: 30px;
-  text-align: center;
-  font-size: 0.9em;
-  color: #555;
+.cancelButton:hover {
+  background-color: #5a6b7b; /* Darker gray */
 }
