--- frontend/src/styles/MySubmission.module.css
@@ -1,47 +1,0 @@
-/* MySubmissions.module.css */
-
-.container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 20px;
-  }
-  
-  .heading {
-    font-size: 2em;
-    margin-bottom: 20px;
-  }
-  
-  .error {
-    color: red;
-    margin-bottom: 20px;
-  }
-  
-  .noArticles {
-    font-size: 1.2em;
-    color: #555;
-  }
-  
-  .table {
-    width: 80%;
-    border-collapse: collapse;
-    margin-top: 20px;
-    border: 1px solid #ddd;
-  }
-  
-  .table th,
-  .table td {
-    border: 1px solid #ddd;
-    padding: 12px 15px;
-    text-align: left;
-  }
-  
-  .table th {
-    background-color: #f2f2f2;
-    font-weight: bold;
-  }
-  
-  .table tr:nth-child(even) {
-    background-color: #f9f9f9;
-  }
-
