+++ frontend/src/styles/SubmitArticlePage.module.css
@@ -1,65 +1,106 @@
-/* SubmitArticlePage.module.css */
-
 .container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-}
-
-.heading {
-  font-size: 2em;
-  margin-bottom: 20px;
+  max-width: 600px;
+  margin: 0 auto;
+  padding: 40px 20px;
+  background-color: #e6f2ff;
+  border-radius: 8px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
 }
 
 .form {
   display: flex;
   flex-direction: column;
-  width: 50%;
-  background: #f9f9f9;
-  padding: 20px;
-  border: 1px solid #ddd;
-  border-radius: 8px;
+}
+
+.heading {
+  text-align: center;
+  margin-bottom: 30px;
+  color: #003366; /* Darker blue for contrast */
 }
 
 .formGroup {
-  margin-bottom: 15px;
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 20px;
 }
 
 .formGroup label {
-  display: block;
-  font-weight: bold;
   margin-bottom: 5px;
+  font-weight: bold;
+  color: #003366; /* Darker blue for consistency */
 }
 
-.formGroup input {
-  width: 100%;
-  padding: 8px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-}
-
-.successMessage {
-  background-color: #e6ffed;
-  border: 1px solid #28a745;
-  color: #155724;
-  border-radius: 4px;
+.formGroup input,
+.formGroup textarea {
   padding: 10px;
-  margin-bottom: 15px; /* Adds spacing below success message */
-  text-align: center;
-  width: 100%;
+  font-size: 16px;
+  border: 1px solid #99ccff; /* Light blue border */
+  border-radius: 5px;
+  background-color: #ffffff; /* White background for inputs */
+}
+
+.formGroup input:focus,
+.formGroup textarea:focus {
+  border-color: #66b3ff; /* Slightly darker blue on focus */
+  outline: none;
+}
+
+.formGroup textarea {
+  min-height: 100px;
+  resize: vertical;
+}
+
+.authorInput {
+  display: flex;
+  align-items: center;
+  margin-bottom: 10px;
+}
+
+.authorInput input {
+  flex: 1;
+}
+
+.authorButtons {
+  display: flex;
+  gap: 5px;
+  margin-left: 10px;
+}
+
+.addButton,
+.removeButton {
+  background-color: #84b8f4; /* Blue color */
+  color: white;
+  border: none;
+  padding: 5px 10px;
+  font-size: 18px;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.removeButton {
+  background-color: #dc3545; /* Red color for remove */
+}
+
+.addButton:hover {
+  background-color: #005bb5;
+}
+
+.removeButton:hover {
+  background-color: #c82333;
 }
 
 .submitButton {
-  background-color: #007bff;
+  padding: 12px 20px;
+  font-size: 18px;
+  background-color: #28a745; /* Green color */
   color: white;
   border: none;
-  padding: 10px 20px;
-  border-radius: 4px;
+  border-radius: 5px;
   cursor: pointer;
-  font-size: 1em;
+  align-self: center;
+  margin-top: 20px;
 }
 
 .submitButton:hover {
-  background-color: #0056b3;
+  background-color: #218838;
 }
