--- frontend/src/styles/Users.module.css
@@ -1,101 +1,0 @@
-/* Users.module.css */
-
-.container {
-  max-width: 1200px;
-  margin: 40px auto;
-  padding: 20px;
-  background-color: #ffffff; /* White background for clarity */
-  border-radius: 8px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  overflow-x: auto; /* Ensures table is scrollable on small devices */
-}
-
-.heading {
-  text-align: center;
-  margin-bottom: 30px;
-  color: #2c3e50; /* Dark blue-gray */
-  font-size: 32px;
-  font-weight: bold;
-}
-
-.error {
-  color: #e74c3c; /* Red color for displaying errors */
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-.table {
-  width: 100%;
-  border-collapse: collapse;
-  margin-top: 20px;
-}
-
-.table th,
-.table td {
-  padding: 12px 15px;
-  border-bottom: 1px solid #ddd;
-  text-align: left;
-}
-
-.table th {
-  background-color: #f2f2f2;
-  color: #2c3e50;
-}
-
-.table tr:nth-child(even) {
-  background-color: #f9f9f9;
-}
-
-.actions {
-  display: flex;
-  gap: 10px;
-  align-items: center; /* Ensures buttons are aligned nicely in the cell */
-}
-
-.editButton,
-.deleteButton { /* Assuming you want a delete button styled similarly */
-  padding: 8px 12px;
-  font-size: 14px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.editButton {
-  background-color: #3498db; /* Blue */
-  color: #fff;
-}
-
-.deleteButton {
-  background-color: #e74c3c; /* Red */
-  color: #fff;
-}
-
-.editButton:hover,
-.deleteButton:hover {
-  opacity: 0.9; /* Slight opacity for hover to give feedback */
-}
-
-/* Responsive Styles */
-@media (max-width: 768px) {
-  .container {
-    padding: 10px;
-  }
-
-  .heading {
-    font-size: 24px; /* Slightly smaller font size for smaller screens */
-  }
-
-  .table th, .table td {
-    padding: 8px; /* Less padding for table headers and cells on smaller screens */
-  }
-
-  .editButton, .deleteButton {
-    padding: 6px 10px; /* Smaller buttons on smaller screens */
-    font-size: 12px;
-  }
-
-  .actions {
-    gap: 5px; /* Less gap between actions buttons on smaller screens */
-  }
-}
