--- frontend/src/styles/ViewAnalysed.module.css
@@ -1,110 +1,0 @@
-/* styles/ViewAnalysed.module.css */
-
-.container {
-    width: 100%;
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 20px;
-    background-color: #f5f5f5;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-  }
-  
-  .header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 20px;
-  }
-  
-  .logo {
-    font-size: 24px;
-    font-weight: bold;
-    margin: 0;
-  }
-  
-  .nav {
-    display: flex;
-    align-items: center;
-    gap: 20px;
-  }
-  
-  .navLink {
-    color: #0070f3;
-    text-decoration: none;
-    font-size: 18px;
-    font-weight: 500;
-    transition: color 0.3s ease;
-  }
-  
-  .navLink:hover {
-    color: #0056b3;
-  }
-  
-  .articlesMenu {
-    position: relative;
-  }
-  
-  .articlesMenu span {
-    cursor: pointer;
-    font-size: 18px;
-    font-weight: 500;
-  }
-  
-  .profileIcon {
-    cursor: pointer;
-    font-size: 20px;
-  }
-  
-  .title {
-    font-size: 22px;
-    margin-bottom: 10px;
-  }
-  
-  .searchContainer {
-    margin-bottom: 20px;
-    text-align: center;
-  }
-  
-  .searchInput {
-    width: 50%;
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-  }
-  
-  .articlesTable {
-    width: 100%;
-    border-collapse: collapse;
-    margin-bottom: 20px;
-  }
-  
-  .articlesTable th,
-  .articlesTable td {
-    padding: 12px;
-    border: 1px solid #ddd;
-    text-align: left;
-  }
-  
-  .articlesTable th {
-    background-color: #f0f0f0;
-    font-weight: bold;
-  }
-  
-  .articlesTable tr:nth-child(even) {
-    background-color: #f9f9f9;
-  }
-  
-  .noArticles {
-    text-align: center;
-    font-style: italic;
-  }
-  
-  .footer {
-    text-align: center;
-    padding: 10px;
-    margin-top: 20px;
-    border-top: 1px solid #ddd;
-  }
-
