+++ frontend/src/styles/analyst.module.css
@@ -1,78 +1,27 @@
-/* Container for the entire form and claim information */
-.formContainer {
-  display: flex;
-  flex-direction: column;
-  padding: 20px;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  background-color: #f9f9f9;
-}
-
-/* Section for displaying article claim information */
-.claimInfo {
-  margin-bottom: 20px;
-  padding: 15px;
-  border: 1px solid #e0e0e0;
-  border-radius: 6px;
-  background-color: #ffffff;
-}
-
-/* Form styling */
-.form {
-  display: flex;
-  flex-direction: column;
-}
-
-/* Form group for label and input elements */
-.formGroup {
-  margin-bottom: 15px;
-}
-
-/* Label styling */
-.formGroup label {
-  font-weight: bold;
-  margin-bottom: 5px;
-}
-
-/* Text area styling for the evidence claim */
-.textAreaInput {
-  width: 100%;
-  height: 100px;
-  padding: 10px;
-  font-size: 14px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  resize: vertical; /* Allow vertical resizing only */
-}
-
-/* Select input styling for evidence result */
-.selectInput {
-  width: 100%;
-  padding: 10px;
-  font-size: 14px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-}
-
-/* Button styling for the update button */
-.submitButton {
-  background-color: #0070f3;
-  color: #ffffff;
-  padding: 10px 15px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 16px;
-  transition: background-color 0.3s;
-}
-
-/* Hover effect for the button */
-.submitButton:hover {
-  background-color: #005bb5;
-}
-
-/* Error message styling */
-.error {
-  color: red;
-  margin-top: 10px;
-}
+/* analyst.module.css */
+.tableContainer {
+    font-family: Arial, sans-serif;
+    margin: 20px;
+    border-collapse: collapse;
+    width: 100%;
+  }
+  
+  .tableHeader {
+    background-color: #f4f4f4;
+    font-weight: bold;
+    text-align: left;
+  }
+  
+  .tableRow {
+    border-bottom: 1px solid #ddd;
+  }
+  
+  .cell {
+    padding: 8px;
+    border-right: 1px solid #ddd;
+  }
+  
+  .lastCell {
+    border-right: none;
+  }
+
