+++ frontend/src/styles/rejected.module.css
@@ -1,88 +1,117 @@
-/* rejected.module.css */
-
+/* styles/rejected.module.css */
 .container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-  background-color: #f9f9f9;
-  min-height: 100vh;
-}
+    max-width: 1000px;
+    margin: 40px auto;
+    padding: 20px;
+    background-color: #f9f9f9;
+    border-radius: 10px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  }
 
-.heading {
-  font-size: 2em;
-  margin-bottom: 20px;
-}
+  .heading {
+    text-align: center;
+    font-size: 28px;
+    color: #2c3e50;
+    margin-bottom: 20px;
+  }
 
-.error {
-  color: red;
-  margin-bottom: 20px;
-}
+  .error {
+    text-align: center;
+    font-size: 16px;
+    color: #e74c3c;
+    margin-bottom: 20px;
+  }
+  .noArticles {
+    text-align: center;
+    font-size: 18px;
+    color: #7f8c8d;
+    margin-top: 20px;
+  }
 
-.noArticles {
-  text-align: center;
-  font-size: 1.2em;
-  color: #555;
-}
+  .table {
+    width: 100%;
+    border-collapse: collapse;
+    margin-top: 20px;
+  }
 
-.table {
-  width: 80%;
-  border-collapse: collapse;
-  margin-top: 20px;
-  border: 1px solid #ddd;
-  background-color: white;
-}
+  .table th,
+  .table td {
+    padding: 12px 15px;
+    border-bottom: 1px solid #ddd;
+    text-align: left;
+  }
 
-.table th,
-.table td {
-  border: 1px solid #ddd;
-  padding: 12px 15px;
-  text-align: left;
-}
+  .table th {
+    background-color: #f2f2f2;
+    color: #2c3e50;
+  }
 
-.table th {
-  background-color: #f2f2f2;
-  font-weight: bold;
-}
+  
+  .table tr:hover {
+    background-color: #f5f5f5;
+  }
+  
+  /* Status Styles */
 
-.table tr:nth-child(even) {
-  background-color: #f9f9f9;
-}
+  .table tr:hover {
+    background-color: #f5f5f5;
+  }
 
-.rejectedStatus {
-  color: red;
-  font-weight: bold;
-}
 
-.noFeedback {
-  color: #999;
-  font-style: italic;
-}
+  .rejectedStatus {
+    color: #e74c3c;
+    font-weight: bold;
+  }
 
-.footer {
-  margin-top: 30px;
-  text-align: center;
-  font-size: 0.9em;
-  color: #555;
-}
 
-.searchBox {
-  width: 60%;
-  padding: 10px;
-  border: 1px solid #ccc;
-  margin-bottom: 20px;
-  font-size: 1em;
-}
+  .approvedStatus {
+    color: #2ecc71;
+    font-weight: bold;
+  }
+  /* Button Styles */
 
-.searchContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-}
+  .buttonApprove,
+  .buttonReject {
+    padding: 8px 12px;
+    margin: 5px;
+    font-size: 14px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+  }
 
-.searchHeader {
-  font-size: 1.2em;
-  font-weight: bold;
-  margin-bottom: 10px;
-}
+  .buttonApprove {
+    background-color: #2ecc71; /* Green */
+    color: white;
+  }
+  
+  .buttonReject {
+    background-color: #e74c3c; /* Red */
+    color: white;
+  }
+  
+  .buttonApprove:hover {
+    background-color: #27ae60;
+  }
+  
+  .buttonReject:hover {
+    background-color: #c0392b;
+  }
+
+  .buttonApprove {
+    background-color: #2ecc71; 
+    color: white;
+  }
+
+  .buttonReject {
+    background-color: #e74c3c; 
+    color: white;
+  }
+
+  .buttonApprove:hover {
+    background-color: #27ae60;
+  }
+
+  .buttonReject:hover {
+    background-color: #c0392b;
+  }
